added compare-trees
