+++ tree-kernel/Makefile
@@ -7,16 +7,20 @@ LIB_SOURCES=\
 	tree-kernel.cpp
 TEST_SOURCES=\
 	test-tree-kernel.cpp
+SOURCES=\
+  compare-trees.cpp
 OBJDIR=.objs
 LIB_OBJECTS=$(addprefix $(OBJDIR)/, $(LIB_SOURCES:.cpp=.o))
 TEST_OBJECTS=$(addprefix $(OBJDIR)/, $(TEST_SOURCES:.cpp=.o))
+OBJECTS=$(addprefix $(OBJDIR)/, $(SOURCES:.cpp=.o))
 LIBSDIR=../tree-parser
 LIBS=tree_parser
 
 TEST_TARGET=test-tree-kernel
 LIB_TARGET=libtree_kernel.a
+TARGET=compare-trees
 
-all: $(LIB_TARGET) $(TEST_TARGET)
+all: $(LIB_TARGET) $(TEST_TARGET) $(TARGET)
 
 $(OBJDIR):
 	mkdir $(OBJDIR)
@@ -33,6 +37,9 @@ $(LIB_TARGET): $(LIB_OBJECTS)
 
 $(TEST_TARGET): $(LIB_OBJECTS) $(TEST_OBJECTS)
 	$(CXX) $(LIB_OBJECTS) $(TEST_OBJECTS) -L$(LIBSDIR) -l$(LIBS) -o $@
+
+$(TARGET): $(LIB_OBJECTS) $(OBJECTS)
+	$(CXX) $(LIB_OBJECTS) $(OBJECTS) -L$(LIBSDIR) -l$(LIBS) -o $@
 
 clean:
 	rm -f $(LIB_TARGET)
